# packages/contracts/http/openapi.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Place HTTP",
    "version": "1.0.0",
    "description": "Place-owned source-only HTTP and browser BFF contract."
  },
  "servers": [],
  "paths": {
...
    "/v1/library/places": {
      "get": {
        "operationId": "listLibraryPlaces",
        "summary": "listLibraryPlaces",
        "security": [
          {
            "placeBearer": []
          }
        ],
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "saved",
                "wanted",
                "rated"
              ],
              "default": "saved"
            }
          },
          {
            "name": "tagIds",
            "in": "query",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "maxItems": 20,
              "uniqueItems": true,
              "items": {
                "type": "string",
                "format": "uuid"
              }
            }
          },
          {
            "name": "tagMatch",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "all",
                "any"
              ],
              "default": "all"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 2048
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return one bounded page of authoritative member Place preferences",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryPlaceListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ProductRequestInvalid"
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "403": {
            "$ref": "#/components/responses/AccessDenied"
          },
          "503": {
            "$ref": "#/components/responses/LibraryQueryUnavailable"
          }
        }
      }
    },
    "/v1/library/collections": {
      "get": {
        "operationId": "listLibraryCollections",
        "summary": "listLibraryCollections",
        "security": [
          {
            "placeBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 2048
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return one bounded page of member Collection summaries",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryCollectionListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ProductRequestInvalid"
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "403": {
            "$ref": "#/components/responses/AccessDenied"
          },
          "503": {
            "$ref": "#/components/responses/LibraryQueryUnavailable"
          }
        }
      }
    },
    "/v1/library/collections/{collectionId}": {
      "parameters": [
        {
          "name": "collectionId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "operationId": "getLibraryCollection",
        "summary": "getLibraryCollection",
        "security": [
          {
            "placeBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 2048
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return Collection metadata and one bounded Place page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryCollectionDetailResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ProductRequestInvalid"
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "403": {
            "$ref": "#/components/responses/AccessDenied"
          },
          "404": {
            "$ref": "#/components/responses/ProductNotFound"
          },
          "503": {
            "$ref": "#/components/responses/LibraryQueryUnavailable"
          }
        }
      }
    },
    "/v1/library/tags": {
      "get": {
        "operationId": "listLibraryTags",
        "summary": "listLibraryTags",
        "security": [
          {
            "placeBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 2048
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return one bounded page of member Tag summaries",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryTagListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ProductRequestInvalid"
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "403": {
            "$ref": "#/components/responses/AccessDenied"
          },
          "503": {
            "$ref": "#/components/responses/LibraryQueryUnavailable"
          }
        }
      }
    },
...
    "/v1/visits": {
      "post": {
        "operationId": "recordPlaceVisit",
        "summary": "recordPlaceVisit",
        "security": [
          {
            "placeBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VisitRecordRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Record an immutable visit occurrence",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisitRecordResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ProductRequestInvalid"
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "403": {
            "$ref": "#/components/responses/AccessDenied"
          },
          "409": {
            "$ref": "#/components/responses/ProductConflict"
          }
        }
      }
    },
...
    "/v1/places/{placeId}/visit-summary": {
      "parameters": [
        {
          "name": "placeId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "operationId": "getPlaceVisitSummary",
        "summary": "getPlaceVisitSummary",
        "security": [
          {
            "placeBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Return the current-member Visit summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisitSummaryResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ProductRequestInvalid"
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "403": {
            "$ref": "#/components/responses/AccessDenied"
          }
        }
      }
    },
    "/v1/places/{placeId}/visits": {
      "parameters": [
        {
          "name": "placeId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "operationId": "listCurrentMemberPlaceVisits",
        "summary": "listCurrentMemberPlaceVisits",
        "security": [
          {
            "placeBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 2048
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return a bounded current-member Visit history",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisitHistoryResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ProductRequestInvalid"
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "403": {
            "$ref": "#/components/responses/AccessDenied"
          },
          "503": {
            "$ref": "#/components/responses/VisitQueryUnavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "placeBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "placeBrowserSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-place_session"
      },
      "placeConnector": {
        "type": "http",
        "scheme": "PlaceConnector"
      }
    },
    "schemas": {
...
      "LibraryPlaceListResponse": {
        "type": "object",
        "properties": {
          "schemaVersion": {
            "type": "string",
            "const": "library-place-list.v2"
          },
          "filter": {
            "type": "object",
            "properties": {
              "state": {
                "type": "string",
                "enum": [
                  "saved",
                  "wanted",
                  "rated"
                ]
              },
              "tagIds": {
                "maxItems": 20,
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                }
              },
              "tagMatch": {
                "type": "string",
                "enum": [
                  "all",
                  "any"
                ]
              }
            },
            "required": [
              "state",
              "tagIds",
              "tagMatch"
            ],
            "additionalProperties": false
          },
          "items": {
            "maxItems": 50,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "placeId": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "saved": {
                  "type": "boolean"
                },
                "wanted": {
                  "type": "boolean"
                },
                "personalRating": {
                  "anyOf": [
                    {
                      "type": "number",
                      "minimum": 0.1,
                      "maximum": 5,
                      "multipleOf": 0.1
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "updatedAt": {
                  "type": "string",
                  "format": "date-time",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                },
                "place": {
                  "anyOf": [
                    {
                      "type": "object",
                      "properties": {
                        "placeId": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                        },
                        "name": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 300
                        },
                        "areaLabel": {
                          "anyOf": [
                            {
                              "type": "string",
                              "minLength": 1,
                              "maxLength": 300
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "location": {
                          "type": "object",
                          "properties": {
                            "latitude": {
                              "type": "number",
                              "minimum": -90,
                              "maximum": 90
                            },
                            "longitude": {
                              "type": "number",
                              "minimum": -180,
                              "maximum": 180
                            }
                          },
                          "required": [
                            "latitude",
                            "longitude"
                          ],
                          "additionalProperties": false
                        },
                        "primaryTaxonomy": {
                          "anyOf": [
                            {
                              "type": "object",
                              "properties": {
                                "key": {
                                  "type": "string",
                                  "minLength": 1,
                                  "maxLength": 128
                                },
                                "label": {
                                  "type": "string",
                                  "minLength": 1,
                                  "maxLength": 160
                                }
                              },
                              "required": [
                                "key",
                                "label"
                              ],
                              "additionalProperties": false
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "taxonomyKeys": {
                          "maxItems": 32,
                          "type": "array",
                          "items": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 128
                          }
                        },
                        "evidence": {
                          "type": "object",
                          "properties": {
                            "status": {
                              "type": "string",
                              "enum": [
                                "verified",
                                "unverified",
                                "conflicted",
                                "stale"
                              ]
                            },
                            "projectedAt": {
                              "type": "string",
                              "format": "date-time",
                              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                            }
                          },
                          "required": [
                            "status",
                            "projectedAt"
                          ],
                          "additionalProperties": false
                        }
                      },
                      "required": [
                        "placeId",
                        "name",
                        "areaLabel",
                        "location",
                        "primaryTaxonomy",
                        "taxonomyKeys",
                        "evidence"
                      ],
                      "additionalProperties": false
                    },
                    {
                      "type": "null"
                    }
                  ]
                }
              },
              "required": [
                "placeId",
                "saved",
                "wanted",
                "personalRating",
                "updatedAt",
                "place"
              ],
              "additionalProperties": false
            }
          },
          "nextCursor": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2048
          }
        },
        "required": [
          "schemaVersion",
          "filter",
          "items"
        ],
        "additionalProperties": false
      },
      "LibraryCollectionListResponse": {
        "type": "object",
        "properties": {
          "schemaVersion": {
            "type": "string",
            "const": "library-collection-list.v1"
          },
          "items": {
            "maxItems": 50,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "collectionId": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "name": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 120
                },
                "description": {
                  "anyOf": [
                    {
                      "type": "string",
                      "maxLength": 2000
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "visibility": {
                  "type": "string",
                  "enum": [
                    "private",
                    "unlisted",
                    "public"
                  ]
                },
                "publicationId": {
                  "anyOf": [
                    {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "placeCount": {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 9007199254740991
                },
                "updatedAt": {
                  "type": "string",
                  "format": "date-time",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                }
              },
              "required": [
                "collectionId",
                "name",
                "description",
                "visibility",
                "publicationId",
                "placeCount",
                "updatedAt"
              ],
              "additionalProperties": false
            }
          },
          "nextCursor": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2048
          }
        },
        "required": [
          "schemaVersion",
          "items"
        ],
        "additionalProperties": false
      },
...
    },
    "responses": {
...
      "ProductRequestInvalid": {
        "description": "The product request is malformed or contains unsupported fields",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "ProductNotFound": {
        "description": "The owned resource or disclosed publication is unavailable",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "LibraryQueryUnavailable": {
        "description": "The bounded Library query is temporarily unavailable",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "AuthenticationRequired": {
        "description": "The bearer token is missing or invalid",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "AccessDenied": {
        "description": "The principal lacks Place access",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
    }
  }
}